kepctl output: extract output logic to pkg/output
Repo seems to be all about CRUD for KEPs. In that context, formatting
output seem like an unrelated concern.

Use an interface/factory pattern to build and use an Output to display
results. Next step: make Query return results, make callers use Output.
